[ST-24] WP: 주문하기 iv. 가게 삭제시 hasSize (2) 가 나오는 버그

+++ streetvendor-api/src/main/java/store/streetvendor/service/store/StoreService.java
@@ -30,7 +30,6 @@ public class StoreService {
 
     @Transactional(readOnly = true)
     public List<StoreResponseDto> getMyStoreList(Long memberId) {
-        // TODO 메뉴랑 결제 방법도 함께 반환해주세요! @yerimkoko
         List<Store> stores = storeRepository.findStoreByBossId(memberId);
         return getStores(stores);
     }
@@ -44,7 +43,10 @@ public class StoreService {
     @Transactional
     public void updateMyStore(Long memberId, Long storeId, StoreUpdateRequest updateRequest) {
         Member member = MemberServiceUtils.findByMemberId(memberRepository, memberId);
-        Store store = StoreServiceUtils.findStoreByStoreIdAndMemberId(storeRepository, storeId, member.getId());
+        Store store = storeRepository.findStoreByStoreIdAndMemberId(storeId, member.getId());
+        if (store == null) {
+            throw new IllegalArgumentException(String.format("해당하는 (%s) 상점이 존재하지 않습니다.", storeId));
+        }
         store.updateStoreInfo(updateRequest.getName(), updateRequest.getDescription(), updateRequest.getPictureUrl(),
             updateRequest.getLocation(), updateRequest.getStartTime(), updateRequest.getEndTime());
         store.updateMenus(updateRequest.getMenus());
@@ -54,8 +56,10 @@ public class StoreService {
     @Transactional
     public void deleteMyStore(Long memberId, Long storeId) {
         Member member = MemberServiceUtils.findByMemberId(memberRepository, memberId);
-        Store store = StoreServiceUtils.findStoreByStoreIdAndMemberId(storeRepository, storeId, member.getId());
+        Store store = storeRepository.findStoreByStoreIdAndMemberId(storeId, member.getId());
+        if (store == null) {
+            throw new IllegalArgumentException(String.format("해당하는 (%s) 상점이 존재하지 않습니다.", storeId));
+        }
         store.delete();
     }
-
 }
